Add support for --upgrade option to automatically upgrade MCU firmware
This option will take the latest firmware from a system data directory,
determine if an upgrade is needed, and do it.

If the upgrade is not possible due to the new application firmware only
supporting the new message passing API and the MCU currently flashed
bootloader firmware only supporting old message passing API, then the
--upgrade option will first flash the version of the firmware that
supports both message passing APIs and only then it will upgrade to the
newest firmware.

This should not be needed because the newest application firmware
supports both message passing APIs.

If the user is upgrading to older application firmware, they will need
to execute
  omnia-mcutool --upgrade
restart the board, and again execute
  omnia-mcutool --upgrade --force
where the --force option is needed so that the bootloader firmware is
also flashed. The user should make sure that this operation is not
interrupted e.g. by power failure (if bootloader flashing operation is
interrupted, the result is a bricked board that will require debug
cables to fix).

If the needed firmware files are not present in
/usr/share/omnia-mcu-firmware directory, then the option will inform the
user where to download the firmwares from.

Signed-off-by: Marek Behún <marek.behun@nic.cz>